Refactor SpecGroup to not subclass Array
